# src/config/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1.0.0",
    "title": "Tech Challenge FIAP",
    "description": "System for order management in a restaurant"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server 🚀"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "General"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/customers/": {
      "post": {
        "tags": [
          "Customers"
        ],
        "description": "Create a new customer",
        "parameters": [
          {
            "name": "customerData",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/customers/cpf/{cpf}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/customers/email/{email}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/products": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "Create a new product",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Get a product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "description": "Update a product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "description": "Delete a product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/products/category/{category}": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Get products by category",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string",
            "enum": [
              "lanche",
              "acompanhamento",
              "bebida",
              "sobremesa"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/orders": {
    "post": {
      "tags": [
        "Order"
      ],
      "description": "Create a new order",
      "parameters": [
        {
          "name": "orderData",
          "in": "body",
          "required": true,
          "schema": {
            "$ref": "#/definitions/OrderInput"
          }
        }
      ],
      "responses": {
        "201": {
          "description": "Created",
          "schema": {
            "type": "object",
            "properties": {
              "order_identifier": {
                "type": "string"
              },
              "orderPaymentQRCode": {
                "type": "string"
              }
            }
          }
        },
        "400": {
          "description": "Bad Request"
        }
      }
    },
    "get": {
      "tags": [
        "Order"
      ],
      "description": "Get all orders",
      "responses": {
        "200": {
          "description": "OK",
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Order"
            }
          }
        },
        "400": {
          "description": "Bad Request"
        }
      }
    }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "Get an order by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Order"
        ],
        "description": "Update an order by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "orderData",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/orders/{id}/qrcode": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "Get order payment QR code",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/orders/{id}/status": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "Get order status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "enum": [
                    "recebido",
                    "preparando",
                    "pronto",
                    "finalizado"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Order"
        ],
        "description": "Update order status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "statusData",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "orderStatus": {
                  "type": "string",
                  "enum": [
                    "recebido",
                    "preparando",
                    "pronto",
                    "finalizado"
                  ]
                },
                "paymentStatus": {
                  "type": "string",
                  "enum": [
                    "pendente",
                    "pago",
                    "cancelado"
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/orders/active": {
      "get": {
        "tags": [
          "Order"
        ],
        "description": "Get all active orders",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "Customer": {
      "type": "object",
      "properties": {
        "cpf": {
          "type": "string",
          "example": "123.456.789-00"
        },
        "email": {
          "type": "string",
          "example": "example@example.com"
        },
        "name": {
          "type": "string",
          "example": "John Doe"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string",
          "example": "123"
        },
        "combo": {
          "type": "object",
          "properties": {
            "LANCHE": {
              "$ref": "#/definitions/Product"
            },
            "ACOMPANHAMENTO": {
              "$ref": "#/definitions/Product"
            },
            "BEBIDA": {
              "$ref": "#/definitions/Product"
            },
            "SOBREMESA": {
              "$ref": "#/definitions/Product"
            }
          }
        },
        "status": {
          "type": "string",
          "enum": [
            "recebido",
            "preparando",
            "pronto",
            "finalizado"
          ],
          "example": "recebido"
        },
        "total": {
          "type": "number",
          "example": 29.99
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "enum": [
            "lanche",
            "acompanhamento",
            "bebida",
            "sobremesa"
          ],
          "example": "lanche"
        },
        "name": {
          "type": "string",
          "example": "Product Name"
        },
        "price": {
          "type": "number",
          "example": 19.99
        },
        "description": {
          "type": "string",
          "example": "This is a product description"
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}